backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
openai==1.65.0
pdfplumber==0.11.5
Pillow==11.1.0
python-dotenv==1.0.1
